% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_fut_irr_ytm.R
\name{bond_fut_irr_ytm}
\alias{bond_fut_irr_ytm}
\title{bond_fut_irr_ytm}
\usage{
bond_fut_irr_ytm(
  bond_call_prices,
  bond_call_strikes,
  bond_put_prices,
  bond_put_strikes,
  stir_call_prices,
  stir_call_strikes,
  stir_put_prices,
  stir_put_strikes,
  r,
  r_2,
  r_3,
  day_count_conv,
  cot_bond,
  bond_fut_price,
  bond_cp,
  bond_cp_f,
  bond_conv_factor,
  sett,
  bond_N = 100,
  bond_matu,
  bond_fut_matu,
  bond_option_matu,
  cot_stir,
  stir_fut_price,
  stir_fut_matu,
  stir_option_matu,
  start_date,
  country = NA,
  currency = NA,
  term_stir
)
}
\arguments{
\item{bond_call_prices}{a vector of call prices on a bond futures, in numeric format}

\item{bond_call_strikes}{a vector of call strikes attached to the call prices, in numeric format}

\item{bond_put_prices}{a vector of put prices on the same bond futures, in numeric format}

\item{bond_put_strikes}{a vector of put strikes attached to the put prices, in numeric format}

\item{stir_call_prices}{a vector of call prices on a STIR futures whose maturity date is near the maturity date of the bond futures (maturity distance below 30 days), in numeric format}

\item{stir_call_strikes}{a vector of call strikes attached to the call prices, in numeric format}

\item{stir_put_prices}{a vector of put prices on the same STIR futures, in numeric format}

\item{stir_put_strikes}{a vector of put strikes attached to the put prices, in numeric format}

\item{r}{a number for the riskfree spot rate whose maturity is equal to the options' maturity, in numeric format}

\item{r_2}{a number for the spot repo funding rate of the underlying bond, with maturity equal to the options' maturity, in numeric format}

\item{r_3}{a number for the spot repo funding rate of the underlying bond, with maturity equal to the futures' maturity, in numeric format}

\item{day_count_conv}{a number for the day count convention, 1 for ACT/ACT, 2 for ACT/360, 3 for ACT/365 and 4 for 30/360, in numeric format}

\item{cot_bond}{a number for the bond options' style, 1 for European options, 2 for American options and 3 for American options with futures-style margin, in numeric format}

\item{bond_fut_price}{a number for the bond futures' price on calibration date, in numeric format}

\item{bond_cp}{a number for the coupon rate of the current Cheapest-to-Deliver Bond, in numeric format}

\item{bond_cp_f}{a number for the frequency of coupon payment of the current Cheapest-to-Deliver Bond, either 1 if the frequency is annual or 2 if semi-annual}

\item{bond_conv_factor}{a number for the conversion factor assigned by the futures exchange to the current Cheapest-to-Deliver Bond of the bond futures, in numeric format}

\item{sett}{a number for the number of days between the ex-coupon date and the coupon payment date of the current Cheapest-to-Deliver Bond, in numeric format}

\item{bond_N}{a number for the value of the principal of the current Cheapest-to-Deliver Bond, in numeric format (100 by default)}

\item{bond_matu}{a date for the maturity date of the current Cheapest-to-Deliver Bond, in Date format}

\item{bond_fut_matu}{a date for the maturity date of the bond futures contract, in Date format}

\item{bond_option_matu}{a date for the maturity date of the bond options, in Date format}

\item{cot_stir}{a number for the STIR options' style, 1 for European options, 2 for American options and 3 for American options with futures-style margin, in numeric format}

\item{stir_fut_price}{a number for the STIR futures' price on calibration date, in numeric format}

\item{stir_fut_matu}{a date for the maturity date of the STIR futures contract, in Date format}

\item{stir_option_matu}{a date for the maturity date of the STIR options, in Date format}

\item{start_date}{a date for the observation date, in Date format}

\item{country}{a character for the country of the issuer of the bond underlying the futures contract, in character format (NA by default)}

\item{currency}{a character for the currency in which the futures contract and the options are traded, in character format (NA by default)}

\item{term_stir}{a number for the term to maturity, in months, of the STIR at observation date, in numeric format}
}
\value{
the parameters (means, standard deviations and weight in the first component lognormal law) of respectively the RND of the bond price and the RND of the bond repo asset price at options' maturity, the mean, standard deviation, skewness, kurtosis, mode, quantiles of order 0.1\%, 0.5\%, 1\%, 5\%, 10\%, 25\%, 50\%, 75\%, 90\%, 95\%, 99\%, 99.5\% and 99.9\% of respectively the bond yield and the bond repo rate distributions in numeric format, the discretized domain and associated RND values of respectively the bond yield to maturity and of the bond repo rate at options' maturity, in numeric format, the RND and CDF plots of respectively the bond yield to maturity and the bond repo rate at options' maturity,  the correlation coefficient between the bond price and the bond repo asset price at options' maturity, the convergence, with 0 indicating successful convergence, in numeric format, the options' prices predicted by the model, in numeric format
}
\description{
bond_fut_irr_ytm
}
\examples{
\donttest{
bond_fut_irr_ytm(c(12.64, 12.14, 11.65, 11.15, 10.65, 10.15, 9.65,
9.16, 8.66, 8.16, 7.67,  7.17, 6.68, 6.19, 5.70, 5.21, 4.73, 4.25,
3.78, 3.32, 2.87, 2.45, 2.05, 1.67, 1.33, 1.03, 0.77, 0.57, 0.41,
0.29, 0.21, 0.15, 0.10, 0.07, 0.06, 0.04, 0.04, 0.03, 0.02, 0.02,
0.02, 0.02, 0.01, 0.01, 0.01, 0.01, 0.01),
c(seq(114, 136.5, 0.5), 137.5),
c(0.01, 0.01, 0.02, 0.02, 0.02, 0.02, 0.02, 0.03, 0.03, 0.03, 0.04,
0.04, 0.05, 0.06, 0.07, 0.08, 0.10, 0.12, 0.15, 0.19, 0.24, 0.32,
0.42, 0.54, 0.70, 0.90, 1.14, 1.44, 1.78, 2.16, 2.58, 3.02, 3.47,
3.94, 4.43, 4.91, 5.41, 5.90, 6.39, 6.89, 7.39, 7.89, 8.38, 8.88,
9.38, 9.88, 10.88),
c(seq(114, 136.5, 0.5), 137.5),
c(1.704999924, 1.579999924, 1.454999924, 1.329999924, 1.204999924,
1.079999924, 0.954999983, 0.892499983, 0.829999983, 0.767499983,
0.704999983, 0.642499983, 0.582499981, 0.519999981, 0.457499981,
0.397499979, 0.334999979, 0.275000006, 0.217500001, 0.162499994,
0.112499997, 0.072499998, 0.039999999, 0.02, 0.0075, 0.0025, 0.0025,
0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0),
c(seq(95.75, 96.5, 0.125), seq(96.5625, 98.75, 0.0625),
seq(98.875, 99.5, 0.125)),
c(0, 0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0,  0.0025,  0.0025,
0.0025,  0.005, 0.005,  0.0075,  0.012499999,  0.02, 0.029999999,
0.055, 0.085000001, 0.127499998, 0.177499995, 0.234999999,
0.297499985, 0.357499987, 0.419999987, 0.482499987, 0.544999957,
0.607499957, 0.669999957, 0.732499957, 0.794999957, 0.857499957,
0.919999957, 0.982499957, 1.044999957, 1.107499957, 1.169999957,
1.232499957, 1.294999957, 1.419999957, 1.544999957, 1.669999957,
1.794999957, 1.919999957, 2.044999838),
c(seq(95.75, 96.5, 0.125), seq(96.5625, 98.75, 0.0625),
seq(98.875, 99.5, 0.125)),
0.0224,
0.0224,
0.02232,
1,
3,
126.66,
0.026,
1,
0.770088,
2,
bond_N = 100,
as.Date("2035-08-15"),
as.Date("2026-09-08"),
as.Date("2026-08-21"),
3,
97.45,
as.Date("2026-09-14"),
as.Date("2026-08-14"),
as.Date("2026-06-17"),
"Germany",
"EUR",
3)
}
}
