% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_yield_spread.R
\name{bond_yield_spread}
\alias{bond_yield_spread}
\title{bond_yield_spread}
\usage{
bond_yield_spread(
  call_prices_1,
  call_strikes_1,
  put_prices_1,
  put_strikes_1,
  call_prices_2,
  call_strikes_2,
  put_prices_2,
  put_strikes_2,
  r,
  r_2,
  day_count_conv_1,
  day_count_conv_2,
  cot_1,
  cot_2,
  ctd_matu_1,
  ctd_matu_2,
  fut_price_1,
  fut_price_2,
  fut_matu,
  option_matu,
  start_date,
  bond_ISIN_1,
  bond_coupon_1,
  bond_cp_f_1,
  bond_matu_1,
  bond_conv_factor_1,
  bond_ytm_1,
  sett_1,
  bond_ISIN_2,
  bond_coupon_2,
  bond_cp_f_2,
  bond_matu_2,
  bond_conv_factor_2,
  bond_ytm_2,
  sett_2,
  bond_nomi = 100,
  corr,
  ctry_1 = NA,
  ctry_2 = NA
)
}
\arguments{
\item{call_prices_1}{a vector of call prices on the first bond futures, in numeric format}

\item{call_strikes_1}{a vector of call strikes attached to the call prices, in numeric format}

\item{put_prices_1}{a vector of put prices on the first bond futures, in numeric format}

\item{put_strikes_1}{a vector of put strikes attached to the put prices, in numeric format}

\item{call_prices_2}{a vector of call prices on the second bond futures, in numeric format}

\item{call_strikes_2}{a vector of call strikes attached to the call prices, in numeric format}

\item{put_prices_2}{a vector of put prices on the second bond futures, in numeric format}

\item{put_strikes_2}{a vector of put strikes attached to the put prices, in numeric format}

\item{r}{a number for the riskfree spot rate whose maturity is equal to the options' maturity, in numeric format}

\item{r_2}{a number for the riskfree spot rate whose maturity is equal to the futures contracts' maturity, in numeric format}

\item{day_count_conv_1}{a number for the day count convention of the first bond futures, 1 for ACT/ACT, 2 for ACT/360, 3 for ACT/365 and 4 for 30/360, in numeric format}

\item{day_count_conv_2}{a number for the day count convention of the second bond futures, 1 for ACT/ACT, 2 for ACT/360, 3 for ACT/365 and 4 for 30/360, in numeric format}

\item{cot_1}{a number for style of the options on the first bond futures, 1 for European options, 2 for American options and 3 for American options with futures-style margin, in numeric format}

\item{cot_2}{a number for style of the options on the second bond futures, 1 for European options, 2 for American options and 3 for American options with futures-style margin, in numeric format}

\item{ctd_matu_1}{a date for the maturity date of the current Cheapest-to-Deliver Bond of the first bond futures, in Date format}

\item{ctd_matu_2}{a date for the maturity date of the current Cheapest-to-Deliver Bond of the second bond futures, in Date format}

\item{fut_price_1}{a number for the first bond futures' price on calibration date, in numeric format}

\item{fut_price_2}{a number for the second bond futures' price on calibration date, in numeric format}

\item{fut_matu}{a date for the bond futures' maturity date, in Date format}

\item{option_matu}{a date for the maturity date of the options, in Date format}

\item{start_date}{a date for the observation date, in Date format}

\item{bond_ISIN_1}{a vector of ISIN codes for the bonds in the delivery basket of the first bond futures, in character format}

\item{bond_coupon_1}{a vector of the corresponding coupon rates for the bonds in the delivery basket, in numeric format}

\item{bond_cp_f_1}{a vector of the corresponding frequencies of coupon payment, either 1 for annual payment or 2 for semiannual payment, in numeric format}

\item{bond_matu_1}{a vector of the corresponding maturity dates for the bonds in the basket of deliverable bonds, in Date format}

\item{bond_conv_factor_1}{a vector of the corresponding conversion factors for the bonds in the delivery basket, in numeric format}

\item{bond_ytm_1}{a vector of the corresponding yield to maturities at observation date for the bonds in the delivery basket, in numeric format}

\item{sett_1}{a number for the number of days between the ex-coupon date and the coupon payment date of the bonds in the delivery basket, in numeric format}

\item{bond_ISIN_2}{a vector of ISIN codes for the bonds in the delivery basket of the second bond futures, in character format}

\item{bond_coupon_2}{a vector of the corresponding coupon rates for the bonds in the delivery basket, in numeric format}

\item{bond_cp_f_2}{a vector of the corresponding frequencies of coupon payment, either 1 for annual payment or 2 for semiannual payment, in numeric format}

\item{bond_matu_2}{a vector of the corresponding maturity dates for the bonds in the basket of deliverable bonds, in Date format}

\item{bond_conv_factor_2}{a vector of the corresponding conversion factors for the bonds in the delivery basket, in numeric format}

\item{bond_ytm_2}{a vector of the corresponding yield to maturities at observation date for the bonds in the delivery basket, in numeric format}

\item{sett_2}{a number for the number of days between the ex-coupon date and the coupon payment date of the bonds in the delivery basket, in numeric format}

\item{bond_nomi}{a single number for the nominal of the bonds (100 by default) in numeric format}

\item{corr}{a number for the correlation coefficient between the returns on the two bond futures prices with identical maturity, in numeric format}

\item{ctry_1}{a character for the nationality of the issuer of the bond in the first bond futures, in character format (NA by default)}

\item{ctry_2}{a character for the nationality of the issuer of the bond in the second bond futures, in character format (NA by default)}
}
\value{
based on the user's correlation coefficient, 10,000 realizations of the bond yield spread between two different issuers or two bonds of the same issuer with different maturities, using options of identical maturity on two bond futures of identical maturity, and a density plot of the spread, in bps
}
\description{
bond_yield_spread
}
\examples{
\donttest{
bond_yield_spread(
c(12.64, 12.14, 11.65, 11.15, 10.65, 10.15,  9.65,  9.16, 8.66,
8.16, 7.67,  7.17, 6.68, 6.19, 5.70, 5.21, 4.73, 4.25, 3.78, 3.32,
2.87,  2.45, 2.05, 1.67, 1.33, 1.03, 0.77, 0.57, 0.41, 0.29,
0.21,  0.15, 0.10, 0.07, 0.06, 0.04, 0.04, 0.03, 0.02, 0.02, 0.02,
0.02,  0.01, 0.01, 0.01, 0.01, 0.01),
c(seq(114, 136.5, 0.5), 137.5),
c(0.01, 0.01, 0.02, 0.02, 0.02, 0.02, 0.02, 0.03, 0.03, 0.03, 0.04, 0.04,
0.05, 0.06, 0.07, 0.08, 0.10, 0.12, 0.15, 0.19, 0.24, 0.32, 0.42, 0.54,
0.70, 0.90, 1.14, 1.44, 1.78, 2.16, 2.58, 3.02, 3.47, 3.94, 4.43, 4.91,
5.41, 5.90, 6.39, 6.89, 7.39, 7.89, 8.38, 8.88, 9.38, 9.88, 10.88),
c(seq(114, 136.5, 0.5), 137.5),
c(13.46, 12.96, 12.46, 11.96, 11.47, 10.97, 10.47, 9.98, 9.48, 8.99, 8.49,
8.00,  7.51, 7.02, 6.53, 6.04, 5.56, 5.08, 4.61, 4.15, 3.69, 3.25, 2.82,
2.41,  2.02, 1.67, 1.34, 1.05, 0.80, 0.59, 0.43, 0.30, 0.21, 0.14, 0.09,
0.06, 0.04,  0.02, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01, 0.01),
seq(105.5, 129, 0.5),
c(0.02, 0.02, 0.02, 0.02, 0.03, 0.03, 0.03, 0.04, 0.04, 0.05, 0.05, 0.06,
0.07, 0.08, 0.09, 0.10, 0.12, 0.14, 0.17, 0.21, 0.25, 0.31, 0.38, 0.47,
0.58, 0.73, 0.90, 1.11, 1.36, 1.65, 1.99, 2.36, 2.77, 3.20, 3.65, 4.12,
4.60, 5.08, 5.57, 6.07, 6.57, 7.06, 7.56, 8.06, 8.56, 9.06, 9.56, 10.06),
seq(105.5, 129, 0.5),
0.0187,
0.019,
1,
1,
3,
3,
as.Date("2035-08-15"),
as.Date("2035-08-01"),
126.66,
119.04,
as.Date("2026-09-08"),
as.Date("2026-08-21"),
as.Date("2026-06-17"),
c("DE000BU2Z056", "DE000BU2Z064"),
c(0.026, 0.029),
rep(1, 2),
as.Date(c("2035-08-15", "2036-02-15")),
c(0.770088, 0.781249),
c(0.02898, 0.02926),
2,
c("IT0005631590", "IT0005648149", "IT0005676504", "IT0005402117",
"IT0005706285", "IT0005433195"),
c(0.0365, 0.0360, 0.0345, 0.0145, 0.0380, 0.0095),
rep(2, 6),
as.Date(c("2035-08-01", "2035-10-01", "2036-02-01", "2036-03-01",
"2036-07-01", "2037-03-01")),
c(0.845212, 0.839556, 0.824415, 0.679587, 0.844354, 0.616457),
c(0.03567, 0.03587, 0.03620, 0.03614, 0.03665, 0.03712),
2,
100,
0.4,
"Germany",
"Italy")
}

}
