% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd_bond_yield.R
\name{ctd_bond_yield}
\alias{ctd_bond_yield}
\title{ctd_bond_yield}
\usage{
ctd_bond_yield(
  call_prices,
  call_strikes,
  put_prices,
  put_strikes,
  nb_log,
  r,
  r_2,
  day_count_conv,
  cot,
  conv_factor,
  ctd_cp,
  ctd_matu,
  cp_f,
  ctd_N,
  sett,
  fut_price,
  fut_matu,
  option_matu,
  start_date,
  nationality = NA,
  currency = NA
)
}
\arguments{
\item{call_prices}{a vector of call prices on a bond futures, in numeric format}

\item{call_strikes}{a vector of call strikes attached to the call prices, in numeric format}

\item{put_prices}{a vector of put prices on the same bond futures, in numeric format}

\item{put_strikes}{a vector of put strikes attached to the put prices, in numeric format}

\item{nb_log}{a number for the number of component densities in the lognormal mixture to model the bond futures' price, either 2 or 3, in numeric format}

\item{r}{a number for the riskfree spot rate whose maturity is equal to the options' maturity, in numeric format}

\item{r_2}{a number for the riskfree spot rate whose maturity is equal to the futures contract's maturity, in numeric format}

\item{day_count_conv}{a number for the day count convention, 1 for ACT/ACT, 2 for ACT/360, 3 for ACT/365 and 4 for 30/360, in numeric format}

\item{cot}{a number for the options' style, 1 for European options, 2 for American options and 3 for American options with futures-style margin, in numeric format}

\item{conv_factor}{a number for the conversion factor assigned by the futures exchange to the current Cheapest-to-Deliver Bond of the bond futures, in numeric format}

\item{ctd_cp}{a number for the coupon rate of the current Cheapest-to-Deliver Bond, in numeric format}

\item{ctd_matu}{a date for the maturity date of the current Cheapest-to-Deliver Bond, in Date format}

\item{cp_f}{a number for the frequency of coupon payment of the current Cheapest-to-Deliver Bond, either 1 if the frequency is annual or 2 if semi-annual}

\item{ctd_N}{a number for the value of the principal of the current Cheapest-to-Deliver Bond, in numeric format}

\item{sett}{a number for the number of days between the ex-coupon date and the coupon payment date of the current Cheapest-to-Deliver Bond, in numeric format}

\item{fut_price}{a number for the bond futures' price on calibration date, in numeric format}

\item{fut_matu}{a date for the maturity date of the bond futures contract, in Date format}

\item{option_matu}{a date for the maturity date of the options, in Date format}

\item{start_date}{a date for the observation date, in Date format}

\item{nationality}{a character for the nationality of the issuer of the bond underlying the futures contract, in character format (NA by default)}

\item{currency}{a character for the currency in which the futures contract and the options are traded, in character format (NA by default)}
}
\value{
A discretized domain of the current Cheapest-to-Deliver bond yield at option's maturity in numeric format, the probability density for each value in the discretized domain in numeric format, the cumulative density for each value in the discretized domain in numeric format, the type of convergence in the optimization in numeric format (0 indicating successful convergence), the mean, standard deviation, skewness and kurtosis of the distribution of the current Cheapest-to-Deliver bond yield in numeric format, quantiles of order 0.1\%, 0.5\%, 1\%, 5\%, 10\%, 25\%, 50\%, 75\%, 90\%, 95\%, 99\%, 99.5\% and 99.9\% of the distribution in numeric format, the mode of the distribution in numeric format, a plot of the RND and a plot of the CDF of the current Cheapest-to-Deliver bond yield
}
\description{
ctd_bond_yield
}
\examples{
\donttest{
ctd_bond_yield(c(10.39,9.92,9.46,9.00,8.55,8.10,7.66,7.23,
6.81,6.39,5.98,5.58,5.20,4.82,4.46,4.10,3.76,3.44,3.13,2.83,2.56,
2.29,2.05,1.82,1.61,1.42,1.25,1.09,0.95,0.82,0.71,0.61,0.53,0.45,
0.38,0.33,0.28,0.23,0.20,0.17,0.14,0.12,0.10,0.08),
seq(106, 127.5, 0.5),
c(0.22,0.25,0.29,0.33,0.38,0.43,0.49,0.56,0.64,0.72,0.81,0.91,
1.03,1.15,1.29, 1.43,1.59,1.77,1.96,2.16,2.39,2.62,2.88,3.15,
3.44,3.75,4.08, 4.42,4.78,5.15,5.54,5.94,6.36,6.78,7.21,7.66,
8.11,8.56,9.03, 9.50,9.97,10.45,10.93,11.41),
seq(106, 127.5, 0.5),
2,
0.0344,
0.035,
1,
3,
0.893,
0.0435,
as.Date("2033-11-01"),
2,
100,
2,
116.17,
as.Date("2024-12-10"),
as.Date("2024-11-22"),
as.Date("2024-06-14"),
"Italian",
"EUR")
}

}
