% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deliv_bonds_charac_bbg.R
\name{deliv_bonds_charac_bbg}
\alias{deliv_bonds_charac_bbg}
\title{deliv_bonds_charac_bbg}
\usage{
deliv_bonds_charac_bbg(bbg_ticker, date)
}
\arguments{
\item{bbg_ticker}{a Bloomberg ticker of a bond futures contract, in character format}

\item{date}{a date for the recovery of the characteristics of the deliverable bonds in the futures contract, in Date format}
}
\value{
For the bonds in the delivery basket, their ISIN code in character format, their coupon rate and frequency of coupon payment in numeric format (1 for annual and 2 for semi-annual), their maturity date in Date format, their conversion factor, yield to maturity and day count convention in numeric format (1 for ACT/ACT, 2 for ACT/360, 3 for ACT/365 and 4 for 30/360). Provided physical delivery of the contract, the same information is also displayed for the current Cheapest-to-Deliver Bond in a second table, with in addition its net basis in currency units of the futures contract, in numeric format
}
\description{
deliv_bonds_charac_bbg
}
\examples{
\dontrun{
deliv_bonds_charac_bbg("UBU6", as.Date("2026-05-19"))
}

}
