% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_ctd_irr.R
\name{proba_ctd_irr}
\alias{proba_ctd_irr}
\title{proba_ctd_irr}
\usage{
proba_ctd_irr(
  bond_call_prices,
  bond_call_strikes,
  bond_put_prices,
  bond_put_strikes,
  stir_call_prices,
  stir_call_strikes,
  stir_put_prices,
  stir_put_strikes,
  r,
  r_2,
  r_3,
  day_count_conv,
  cot_bond,
  bond_fut_price,
  bond_cp,
  bond_cp_f,
  bond_conv_factor,
  bond_ytm,
  sett,
  Nomi = 100,
  bond_matu,
  bond_ISIN,
  bond_fut_matu,
  bond_option_matu,
  cot_stir,
  stir_fut_price,
  stir_fut_matu,
  stir_option_matu,
  start_date,
  term_stir
)
}
\arguments{
\item{bond_call_prices}{a vector of call prices on a bond futures, in numeric format}

\item{bond_call_strikes}{a vector of call strikes attached to the call prices, in numeric format}

\item{bond_put_prices}{a vector of put prices on the same bond futures, in numeric format}

\item{bond_put_strikes}{a vector of put strikes attached to the put prices, in numeric format}

\item{stir_call_prices}{a vector of call prices on a STIR futures whose maturity date is near the maturity date of the bond futures (maturity distance below 30 days), in numeric format}

\item{stir_call_strikes}{a vector of call strikes attached to the call prices, in numeric format}

\item{stir_put_prices}{a vector of put prices on the same STIR futures, in numeric format}

\item{stir_put_strikes}{a vector of put strikes attached to the put prices, in numeric format}

\item{r}{a number for the riskfree spot rate whose maturity is equal to the options' maturity, in numeric format}

\item{r_2}{a number for the spot repo funding rate of the underlying bond, with maturity equal to the options' maturity, in numeric format}

\item{r_3}{a number for the spot repo funding rate of the underlying bond, with maturity equal to the futures' maturity, in numeric format}

\item{day_count_conv}{a number for the day count convention, 1 for ACT/ACT, 2 for ACT/360, 3 for ACT/365 and 4 for 30/360, in numeric format}

\item{cot_bond}{a number for the bond options' style, 1 for European options, 2 for American options and 3 for American options with futures-style margin, in numeric format}

\item{bond_fut_price}{a number for the bond futures' price on calibration date, in numeric format}

\item{bond_cp}{a vector of the coupon rates for the bonds in the delivery basket, in numeric format}

\item{bond_cp_f}{a vector of the corresponding frequencies of coupon payment, either 1 for annual payment or 2 for semiannual payment, in numeric format}

\item{bond_conv_factor}{a vector of the corresponding conversion factors for the bonds in the delivery basket, in numeric format}

\item{bond_ytm}{a vector of the corresponding yields to maturity at observation date for the bonds in the delivery basket, in numeric format}

\item{sett}{a number for the number of days between the ex-coupon date and the coupon payment date of the current Cheapest-to-Deliver Bond, in numeric format}

\item{Nomi}{a single number for the value of the principal of the bonds in the delivery basket, in numeric format (100 by default)}

\item{bond_matu}{a vector of the corresponding maturity dates for the bonds in the basket of deliverable bonds, in Date format}

\item{bond_ISIN}{a vector of the corresponding ISIN codes for the bonds in the delivery basket of the bond futures, in character format}

\item{bond_fut_matu}{a date for the maturity date of the bond futures contract, in Date format}

\item{bond_option_matu}{a date for the maturity date of the bond options, in Date format}

\item{cot_stir}{a number for the STIR options' style, 1 for European options, 2 for American options and 3 for American options with futures-style margin, in numeric format}

\item{stir_fut_price}{a number for the STIR futures' price on calibration date, in numeric format}

\item{stir_fut_matu}{a date for the maturity date of the STIR futures contract, in Date format}

\item{stir_option_matu}{a date for the maturity date of the STIR options, in Date format}

\item{start_date}{a date for the observation date, in Date format}

\item{term_stir}{a number for the term to maturity, in months, of the STIR at observation date, in numeric format}
}
\value{
for the bonds in the delivery basket, their ISIN in character format and their probability of being the CtD bond at options' maturity based on each deliverable bond's joint distribution of bond yield and repo rate, in numeric format
}
\description{
proba_ctd_irr
}
\examples{
\donttest{
proba_ctd_irr(
c(19.12, 18.12, 17.12, 16.14, 15.14, 14.16, 13.18, 12.20, 11.22, 10.24, 9.30, 8.34,
7.42, 6.52, 5.66, 4.84, 4.06, 3.32, 2.66, 2.08, 1.64, 1.28, 1.00, 0.78, 0.60, 0.46,
0.34, 0.26, 0.20, 0.14, 0.10, 0.08, 0.06, 0.04, 0.04, 0.02, 0.02, 0.02, 0.02, 0.02,
0.02, 0.02, 0.02, 0.02, 0.02, 0.02, 0.02, 0.02),
seq(84, 131),
c(0.04, 0.04, 0.04, 0.06, 0.06, 0.08, 0.10, 0.12, 0.14, 0.16, 0.22, 0.26, 0.34, 0.44,
0.58, 0.76, 0.98, 1.24, 1.58, 2.00, 2.56, 3.20, 3.92, 4.70, 5.52, 6.38, 7.26, 8.18,
9.12, 10.06, 11.02, 12.00, 12.98, 13.96, 14.96, 15.94, 16.94, 17.94, 18.92, 19.92,
20.92, 21.92, 22.92, 23.92, 24.92, 25.92, 26.92, 27.92),
seq(84, 131),
c(3.1850, 3.0600, 2.9350, 2.8100, 2.6850, 2.5600, 2.4350, 2.3100, 2.1850, 2.0625,
1.9375, 1.8125, 1.6875, 1.5625, 1.4375, 1.3150, 1.1900, 1.1275, 1.0650, 0.9400,
0.8800, 0.8175, 0.7550, 0.6950, 0.6325, 0.5725, 0.5125, 0.4525, 0.3925, 0.3350,
0.2775, 0.2225, 0.1700, 0.1250, 0.0875, 0.0575, 0.0325, 0.0150, 0.0075, 0.0050,
0.0025, 0.0025, 0.0025, 0.0025, 0.0025, 0.0025, 0.0025, 0.0025, 0.0025, 0.0025,
0.0025, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000,
0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000,
0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000),
c(seq(94, 96, 0.125), 96.0625, 96.125, seq(96.25, 99.25, 0.0625), 99.3750,
99.5000, 99.6250,  99.6875, 99.7500,  99.8750, 100.0000, 101.0000),
c(0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0000, 0.0025,
0.0025, 0.0025, 0.0025, 0.0025, 0.0025, 0.0050, 0.0050, 0.0050, 0.0050, 0.0050,
0.0075, 0.0075, 0.0075, 0.0100, 0.0100, 0.0125, 0.0150, 0.0175, 0.0200, 0.0250,
0.0300, 0.0375, 0.0475, 0.0650, 0.0900, 0.1225, 0.1600, 0.2050, 0.2600, 0.3200,
0.3800, 0.4425, 0.5050, 0.5675, 0.6300, 0.6925, 0.7550, 0.8175, 0.8800, 0.9425,
1.0050, 1.0650, 1.1275, 1.1900, 1.2525, 1.3150, 1.3775, 1.4400, 1.5025, 1.5650,
1.6275, 1.6900, 1.7525, 1.8150, 1.8775, 1.9400, 2.0025, 2.0650, 2.1900, 2.3150,
2.4400, 2.5025, 2.5650, 2.6900, 2.8150, 3.8150),
c(seq(94, 96, 0.125), 96.0625, 96.125, seq(96.25, 99.25, 0.0625), 99.3750,
99.5000, 99.6250,  99.6875, 99.7500,  99.8750, 100.0000, 101.0000),
0.02303,
0.02303,
0.02388,
1,
3,
103.08,
c(0.000, 0.018, 0.018, 0.025, 0.029),
rep(1, 5),
c(0.365252, 0.643086, 0.643086, 0.751530, 0.810737),
c(0.03811, 0.03816, 0.03806, 0.03813, 0.03812),
2,
Nomi = 100,
as.Date(c("2052-08-15", "2053-08-15", "2053-08-15", "2054-08-15", "2056-08-15")),
c("DE0001102572", "DE0001102614", "DE0001030757", "DE000BU2D004", "DE000BU2D012"),
as.Date("2026-12-08"),
as.Date("2026-11-20"),
3,
97.185,
as.Date("2026-12-14"),
as.Date("2026-12-14"),
as.Date("2026-08-31"),
3)
}

}
