% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proba_ctd_opt.R
\name{proba_ctd_opt}
\alias{proba_ctd_opt}
\title{proba_ctd_opt}
\usage{
proba_ctd_opt(
  call_prices,
  call_strikes,
  put_prices,
  put_strikes,
  nb_log,
  r,
  r_2,
  day_count_conv,
  cot,
  ctd_matu,
  fut_price,
  fut_matu,
  option_matu,
  start_date,
  bond_ISIN,
  bond_coupon,
  bond_cp_f,
  bond_matu,
  bond_nomi = 100,
  bond_conv_factor,
  bond_ytm,
  sett
)
}
\arguments{
\item{call_prices}{a vector of call prices on a bond futures, in numeric format}

\item{call_strikes}{a vector of call strikes attached to the call prices, in numeric format}

\item{put_prices}{a vector of put prices on the same bond futures, in numeric format}

\item{put_strikes}{a vector of put strikes attached to the put prices, in numeric format}

\item{nb_log}{a number for the number of component densities in the lognormal mixture to model the bond futures' price, either 2 or 3, in numeric format}

\item{r}{a number for the riskfree spot rate whose maturity is equal to the options' maturity, in numeric format}

\item{r_2}{a number for the riskfree spot rate whose maturity is equal to the futures' maturity, in numeric format}

\item{day_count_conv}{a number for the day count convention, 1 for ACT/ACT, 2 for ACT/360, 3 for ACT/365 and 4 for 30/360, in numeric format}

\item{cot}{a number for the options' style, 1 for European options, 2 for American options and 3 for American options with futures-style margin, in numeric format}

\item{ctd_matu}{a date for the maturity date of the current Cheapest-to-Deliver Bond of the bond futures, in Date format}

\item{fut_price}{a number for the bond futures' price on calibration date, in numeric format}

\item{fut_matu}{a date for the bond futures' maturity date, in Date format}

\item{option_matu}{a date for the maturity date of the options, in Date format}

\item{start_date}{a date for the observation date, in Date format}

\item{bond_ISIN}{a vector of ISIN codes for the bonds in the delivery basket of the bond futures, in character format}

\item{bond_coupon}{a vector of the corresponding coupon rates for the bonds in the delivery basket, in numeric format}

\item{bond_cp_f}{a vector of the corresponding frequencies of coupon payment, either 1 for annual payment or 2 for semiannual payment, in numeric format}

\item{bond_matu}{a vector of the corresponding maturity dates for the bonds in the basket of deliverable bonds, in Date format}

\item{bond_nomi}{a single number for the nominal of the bonds (100 by default) in numeric format}

\item{bond_conv_factor}{a vector of the corresponding conversion factors for the bonds in the delivery basket, in numeric format}

\item{bond_ytm}{a vector of the corresponding yields to maturity at observation date for the bonds in the delivery basket, in numeric format}

\item{sett}{a number for the number of days between the ex-coupon date and the coupon payment date of the bonds in the delivery basket, in numeric format}
}
\value{
for the bonds in the delivery basket, their ISIN in character format and their probability of being the CtD bond at options' maturity relying successively on their net compared bases at options' maturity and on their compared implied repo rates at option's maturity, in numeric format
}
\description{
proba_ctd_opt
}
\examples{
\donttest{
proba_ctd_opt( c(24.10, 23.10, 22.12, 21.12, 20.12, 19.14, 18.14, 17.16, 16.18, 15.20,
14.22, 13.24, 12.28, 11.32, 10.36, 9.44, 8.50, 7.60, 6.72, 5.86, 5.04, 4.28,
3.56, 2.88, 2.30, 1.78,  1.36, 1.02, 0.76, 0.56, 0.42, 0.30, 0.22, 0.18, 0.14,
0.10,  0.08, 0.06, 0.06, 0.04, 0.04, 0.02, 0.02, 0.02, 0.02, 0.02, 0.02),
seq(85, 131),
c(0.02,  0.02, 0.02, 0.02, 0.02, 0.04, 0.04, 0.06, 0.08, 0.10, 0.12, 0.14, 0.18, 0.22,
0.26, 0.34,  0.40, 0.50, 0.62, 0.76, 0.94, 1.18, 1.46, 1.78, 2.20, 2.68, 3.26, 3.92,
4.66,  5.46, 6.32, 7.20, 8.12, 9.08, 10.04, 11.00, 11.98, 12.96, 13.96, 14.94, 15.94,
16.92, 17.92, 18.92, 19.92, 20.92, 21.92),
seq(85, 131),
2,
0.0187,
0.019,
1,
3,
as.Date("2054-08-15"),
109.1,
as.Date("2026-09-08"),
as.Date("2026-08-21"),
as.Date("2026-05-28"),
c("DE0001102572", "DE0001102614", "DE0001030757", "DE000BU2D004", "DE000BU2D012"),
c(0.000, 0.018, 0.018, 0.025, 0.029),
rep(1, 5),
as.Date(c("2052-08-15", "2053-08-15", "2053-08-15", "2054-08-15", "2056-08-15")),
100,
c(0.361698, 0.641260, 0.641260, 0.750372, 0.809987),
c(0.03500, 0.03507, 0.03492, 0.03510, 0.03514),
2)
}

}
