% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stir_future_charac_bbg.R
\name{stir_future_charac_bbg}
\alias{stir_future_charac_bbg}
\title{stir_future_charac_bbg}
\usage{
stir_future_charac_bbg(bbg_tickers)
}
\arguments{
\item{bbg_tickers}{a vector of Bloomberg tickers of STIR futures contracts, in character format}
}
\value{
for each STIR futures contract, its maturity date in Date format, its currency, its listing place and its underlying asset all in character format
}
\description{
stir_future_charac_bbg
}
\examples{
\dontrun{
stir_future_charac_bbg(c("FFU24", "ERM6"))
}
}
