% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bond_future_charac_bbg.R
\name{bond_future_charac_bbg}
\alias{bond_future_charac_bbg}
\title{bond_future_charac_bbg}
\usage{
bond_future_charac_bbg(bbg_tickers)
}
\arguments{
\item{bbg_tickers}{a vector of Bloomberg tickers of bond futures contracts, in character format}
}
\value{
for each bond future contract, its maturity date in Date format, its currency, its listing place and its underlying asset in character format. Provided the contract has a Cheapest-to-Deliver Bond, its ISIN code in character format, its conversion factor in numeric format, its maturity date in Date format, its (annualized) coupon rate in numeric format, its frequency of coupon payment and its day count convention in character format
}
\description{
bond_future_charac_bbg
}
\examples{
\dontrun{
bond_future_charac_bbg(c("TYU24", "CNM6", "KAAH6", "IKM6", "OATU6"))
}
}
