% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_prices_bbg.R
\name{option_prices_bbg}
\alias{option_prices_bbg}
\title{option_prices_bbg}
\usage{
option_prices_bbg(fut_bbg_ticker, date)
}
\arguments{
\item{fut_bbg_ticker}{a single Bloomberg ticker of a STIR or bond futures contract, in character format}

\item{date}{a date for the recovery of options' closing prices, in Date format}
}
\value{
provided there are listed options on the futures, call and put options strike prices and attached call and put options closing market prices in numeric format, the maturity date of the options in Date format, the option's style in character format, the option's currency in character format, the listing place of the options in character format, the closing price of the futures contract in numeric format. When options of different maturities are listed on the same futures contract, this information concerns options whose maturity date is closest to the maturity date of the futures contract.
}
\description{
option_prices_bbg
}
\examples{
\dontrun{
option_prices_bbg("ERM6", as.Date("2026-03-20"))
}

}
