% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/option_prices_bbg.R
\name{option_prices_bbg}
\alias{option_prices_bbg}
\title{option_prices_bbg}
\usage{
option_prices_bbg(fut_bbg_ticker, date)
}
\arguments{
\item{fut_bbg_ticker}{a Bloomberg ticker of a future contract, in character format}

\item{date}{a date for the recovery of options' closing prices, in Date format}
}
\value{
provided options on the futures contract exist, call and put options closing market prices by strike prices, attached call and put options strike prices, in numeric format, the maturity date of the options, in Date format, the option style, in character format, the option currency, in character format, the closing price of the futures contract, in numeric format. When options of different maturities are listed on the same futures contract, returns prices and characteristics of options whose maturity date is closest to the maturity date of the futures contract.
}
\description{
option_prices_bbg
}
\examples{
\dontrun{
option_prices_bbg("ERM6", as.Date("2026-03-20"))
}

}
